chore(scan): fix 2 aislop warnings on 1.21 branch
- Moved task(), tryRestoreFromMojang(), MojangRestoreResult,
  deriveReason(), getLocalizedString() to new SkinRefreshHandler.java
  (fixes complexity/file-too-large: SkinCommand 457→379 lines)
- Extracted fetchSkinProperty() and handleSkinCompletion() from
  skinAction() method (fixes complexity/function-too-long:
  skinAction 124→40 lines)
- Updated SkinCommandTest.java references to SkinRefreshHandler
